Various changes and fixes
- Changed SZArrays to more closely match .Net Arrays
- Fixed generic method processing corruption
- Re-added null pointer exception page
- All local variables now initially reside on the stack to stop bugs
from occurring
- Simplified code according to VS2015 recommendations

+++ Source/Korlib/System/Char.cs
@@ -53,7 +53,7 @@ namespace System
 
 		public override int GetHashCode()
 		{
-			return (int)_value;
+			return _value;
 		}
 
 		private static bool IsWhiteSpaceLatin1(char c)
