+++ Source/Mosa.Test.Collection.x86.xUnit/UInt16Fixture.cs
@@ -44,7 +44,7 @@ namespace Mosa.Test.Collection.x86.xUnit
 		//[ExpectedException(typeof(DivideByZeroException))]
 		public void DivU2U2DivideByZeroException(ushort a)
 		{
-			Assert.Equal(UInt16Tests.DivU2U2(a, (ushort)0), Run<int>("Mosa.Test.Collection.UInt16Tests.DivU2U2", a, (ushort)0));
+			Assert.Equal(UInt16Tests.DivU2U2(a, 0), Run<int>("Mosa.Test.Collection.UInt16Tests.DivU2U2", a, (ushort)0));
 		}
 
 		[Theory]
@@ -63,7 +63,7 @@ namespace Mosa.Test.Collection.x86.xUnit
 		//[ExpectedException(typeof(DivideByZeroException))]
 		public void RemU2U2DivideByZeroException(ushort a)
 		{
-			Assert.Equal(UInt16Tests.RemU2U2(a, (ushort)0), Run<int>("Mosa.Test.Collection.UInt16Tests.RemU2U2", a, (ushort)0));
+			Assert.Equal(UInt16Tests.RemU2U2(a, 0), Run<int>("Mosa.Test.Collection.UInt16Tests.RemU2U2", a, (ushort)0));
 		}
 
 		[Theory]
