+++ Source/Mosa.TinyCPUSimulator.Adaptor/SimLinker.cs
@@ -87,7 +87,7 @@ namespace Mosa.TinyCPUSimulator.Adaptor
 
 			foreach (var symbol in Symbols)
 			{
-				simAdapter.SimCPU.SetSymbol(symbol.Name, symbol.VirtualAddress, (ulong)symbol.Size);
+				simAdapter.SimCPU.SetSymbol(symbol.Name, symbol.VirtualAddress, symbol.Size);
 			}
 
 			foreach (var symbol in symbolData)
