+++ Source/Mosa.TinyCPUSimulator.x86/Opcodes/Cmp.cs
@@ -11,8 +11,8 @@ namespace Mosa.TinyCPUSimulator.x86.Opcodes
 			uint b = LoadValue(cpu, instruction.Operand2);
 			int size = instruction.Operand2.Size;
 
-			long s = (long)(int)a - (long)(int)b;
-			ulong u = (ulong)a - (ulong)b;
+			long s = (int)a - (long)(int)b;
+			ulong u = a - (ulong)b;
 
 			if (size == 32)
 				UpdateFlags(cpu, size, s, u, true, true, true, true, true);
